- 3.1.0_beta3:   - Add support for up to 3 "Profiles", which can be used to quickly switch between different brew setings (eg. when different beans are used).     - An individual profile comprises of following settings:       - Brew Temperature / setPoint       - Brew Time       - Preinfusion time and pause       - Coldstart Temperature / startTemp     - Profiles can be selected by any means, eg mqtt/blynk/menu/.. . The variable for this is "profile".     - The selected profile is also shown on the display, see new setting ENABLE_PROFILE_STATUS.     - ATTENTION: Following variable names has changed and therefore any mqtt dashboards and tools have to be adapted (eg IotMQTT/Grafana/nodered):       - setPoint -> activeSetPoint       - brewtime -> brewTime       - starttemp -> activeStartTemp       - preinfusion -> activePreinfusion       - preinfusionpause -> activePreinfusionPause       - Hint: Changing the value of any of variables (eg using mqtt/blynk) will update the respective setting of the currently activated profile.     - Blynk:       - Added new virtual Pin V3 for "profile", which can be any number between 1 and 3.       - Added [Blynk Showcase](https://github.com/medlor/bleeding-edge-ranciliopid/tree/master/blynk/blynk-showcase_v3.1.0.mp4).   - userConfig.h Changes:     - Added:       - ENABLE_PROFILE_STATUS       - Renamed SETPOINT to SETPOINT1       - Renamed BREWTIME to SETPOINT1       - Renamed PREINFUSION to PREINFUSION1       - Renamed PREINFUSION_PAUSE to PREINFUSION_PAUSE1     - Added:       - SETPOINT2, SETPOINT3       - STARTTEMP2, STARTTEMP3       - PREINFUSION2, PREINFUSION3       - PREINFUSION_PAUSE2, PREINFUSION_PAUSE3   - Updated IoTMQTT Dashboard [IoTMQTTPanel-perfect_coffee_pid_v3.1.0_v1.json](https://github.com/medlor/bleeding-edge-ranciliopid/tree/3.1.0_beta/IoTMQTTPanel/IoTMQTTPanel-perfect_coffee_pid_v3.1.0_v1.json) to include Profile and Action support. [Showcase](https://github.com/medlor/bleeding-edge-ranciliopid/tree/master/IoTMQTTPanel/IoTMQTT-showcase_v3.1.0.mp4).   - Refactor eeprom stuff to extra file.   - Refactor all variables from double to float.   - Make blynk optional on compilation.   - Fixes:     - Fix blynk "heater output" not being shown.

+++ rancilio-pid/rancilio-pid.h
@@ -74,12 +74,12 @@ extern RemoteDebug Debug;
 #define ALIGN_LEFT 0
 
 // returns heater utilization in percent
-double convertOutputToUtilisation(double);
+float convertOutputToUtilisation(double);
 // returns heater utilization in Output
-double convertUtilisationToOutput(double);
-double pastTemperatureChange(int);
-double pastTemperatureChange(int, bool);
-double getCurrentTemperature();
+double convertUtilisationToOutput(float);
+float pastTemperatureChange(int);
+float pastTemperatureChange(int, bool);
+float getCurrentTemperature();
 float readTemperatureFromSensor();
 bool almostEqual(float, float);
 void print_settings();
